backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
pydantic==2.5.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
prophet==1.1.4
joblib==1.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 